Moves GcodeHelper to PreviewHandlerCommon
