+++ src/modules/previewpane/common/Utilities/GcodeThumbnailFormat.cs
@@ -2,7 +2,7 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace GcodeFileFormat
+namespace Common.Utilities
 {
     /// <summary>
     /// The gcode thumbnail image format.
